Moved popup into view folder
