+++ src/view/popup/__tests__/mockChrome.ts
