+++ src/view/popup/__tests__/usePopupHistory.test.ts
@@ -1,7 +1,7 @@
 import { renderHook, act } from "@testing-library/react";
 
 import { usePopup } from "../usePopup";
-import { ICalculationsRepository } from "../data/calculations/ICalculationsRepository";
+import { ICalculationsRepository } from "../../../data/calculations/ICalculationsRepository";
 import { mockChrome } from "./mockChrome";
 
 (global as any).chrome = mockChrome;
