+++ src/view/popup/__tests__/usePopupSelectedCountries.test.ts
@@ -1,7 +1,7 @@
 import { renderHook, act } from "@testing-library/react";
 
-import { CountryName } from "../data/constants/CountryEmissions";
-import { areMapsDeepEqual } from "../utils/helpers/areMapsDeepEqual";
+import { CountryName } from "../../../data/constants/CountryEmissions";
+import { areMapsDeepEqual } from "../../../utils/helpers/areMapsDeepEqual";
 import { usePopup } from "../usePopup";
 import { mockChrome } from "./mockChrome";
