+++ src/view/popup/popup.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 import { createRoot } from "react-dom/client";
-import { CountryDropdown } from "./components/CountryDropdown";
+import { CountryDropdown } from "../../components/CountryDropdown";
 import { usePopup } from "./usePopup";
-import { SelectedCountries } from "./components/selected-countries/SelectedCountries";
-import { CalculationHistory } from "./components/calculation-history/CalculationHistory";
-import { formatBytesString } from "./utils/helpers/formatBytesString";
+import { SelectedCountries } from "../../components/selected-countries/SelectedCountries";
+import { CalculationHistory } from "../../components/calculation-history/CalculationHistory";
+import { formatBytesString } from "../../utils/helpers/formatBytesString";
 
 export const Popup = () => {
     const {
